onclite: Restore better load balancing on boot
As it turned out, allowing background, system-background, restricted,
foreground and top-app to use all 8 cores results in severe jitter and lag
whenever background tasks are running, whether it be on boot, or during waking up
from sleep.

This commit readds pinning background to cores 0-1, system-background and
restricted to cores 0-3, foreground to cores 0-6 and top-app to cores 0-7.

Change-Id: Idb76879e75e05b437600f0b689c260c674eb7eea

Signed-off-by: Aryan Karan <aryankaran28022004@gmail.com>